Refactor: update name of CommunicationChannelInterface
