--- src/main/java/CommunicationChannelInterface.java
@@ -1,7 +1,0 @@
-import java.io.InputStream;
-import java.io.OutputStream;
-
-public interface CommunicationChannelInterface {
-    OutputStream getOutputStream();
-    InputStream getInputStream();
-}
