+++ src/main/java/SocketManager.java
@@ -3,5 +3,5 @@ import java.io.IOException;
 public interface SocketManager {
     void connect() throws IOException;
     void listen(int port) throws IOException;
-    CommunicationChannelInterface communicationChannel();
+    CommunicationChannel communicationChannel();
 }
